+++ lib/src/utilities/converter.dart
@@ -25,6 +25,7 @@ import 'package:cornerstone/src/core/failure.dart';
 /// }
 /// ```
 abstract class ConvertToFailure<Type> {
+  const ConvertToFailure();
   Failure<Type> call(dynamic e);
 }
 
@@ -32,5 +33,6 @@ abstract class ConvertToFailure<Type> {
 /// Snapshot from data. Used in utility repos such as
 /// [CornerstonePersistentRepositoryMixin].
 abstract class ConvertToSnapshot<Snap> {
+  const ConvertToSnapshot();
   Snap call(Map<String, dynamic> data);
 }